feat: Amend bringIntoView Modifier extension to be compatibel with nay ScrollableState type
